refactored ot databricks function
